Update project
Go to CHANGELOG.md for changelog. Thank you!

+++ src/netcore3.1/Objects/SubjectFee.cs
@@ -20,7 +20,7 @@ namespace ZoeMeow.DUTAPI.Objects
         /// <summary>
         /// Subject credit
         /// </summary>
-        public int Credit { get; set; } = 0;
+        public float Credit { get; set; } = 0;
         #endregion
 
         #region Subject Information
